requirements_test.txt
pytest-django==2.8.0
